requirements.txt
flask==1.0.1
gevent==1.4.0
uWSGI==2.0.18
voluptuous==0.11.5
pytest-flask==0.15.0
requests==2.22.0
attrdict==2.0.1